Add RimDepthTrackPlot base class for RimWellLogPlot.
